Refactor the auth utils
